Update text fixtures
The update to `pandas` resulted in some assertion failures due to
our use of the index as the eventual id value in quiz_attempts.
The version update had a slightly different behavior in ordering of
rows after a merge, which superficially impacted two outputs. For
this iteration I'm just updating the fixtures versus trying to modify
the assertions to be less brittle. Maybe we'll revisit that decision
in the future =).